Added note about supported Draft versions
When I got started working with this library I got frustrated when my Schema didn't work as expected. It wasn't until I looked at the test files that I realized I was using features of a Draft that was not supported by this library. A note like the one I added would have saved me a lot of time trying to debug my Schema. So I am adding it now for everyone that comes after me.